+++ fvm/context.go
@@ -85,9 +85,8 @@ func defaultContext(logger zerolog.Logger) Context {
 		TransactionProcessors: []TransactionProcessor{
 			NewTransactionSignatureVerifier(AccountKeyWeightThreshold),
 			NewTransactionSequenceNumberChecker(),
-			NewTransactionFeeDeductor(),
 			NewTransactionInvocator(logger),
-			NewTransactionStorageLimiter(),
+			NewTransactionFeeDeductor(),
 		},
 		ScriptProcessors: []ScriptProcessor{
 			NewScriptInvocator(),
